[glean-lsp] Account for local changes
If the source file has been changed relative to the version that was
indexed, we would like the LSP to update its results to account for the
local changes so that we can still get meaningful IDE features using the
Glean data even for locally-modified files.

This is a basic implementation of that idea. We diff the file contents
and then apply a line offset to symbol locations, or omit results that
have been deleted locally. This depends on the indexer having stored the
full file content (e.g. the Haskell indexer supports this when the
`--store-src` option is given). Only a line offset is supported for now;
edited lines are considered deleted from the original file.

This is only approximately correct in general: every edit to a file
potentially induces arbitrary semantic changes, so just applying a line
offset won't always be right. But it's often the right thing and in
practice it's terribly useful.

Furthermore there are limitations to what we can do with reasonable
performance. If you perform a find-references request that returns
results in 100+ files, we're not going to open all of those files and
diff the source contents to find out whether the symbols still exist, so
we only do this offsetting for files that are currently open.